feat(capture-rs): refactor quota limiters + add faceted quota event drop metrics (#37325)
